requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
tavily-python==0.3.1
python-multipart==0.0.6
